server/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
psycopg2-binary==2.9.7
pandas==2.2.3
scikit-learn==1.3.2
numpy==1.26.4
python-dotenv==1.0.0
Werkzeug==2.2.3